pyproject.toml
[project]
name = "project1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-generativeai>=0.8.6",
    "pypdf2>=3.0.1",
    "python-dotenv>=1.2.1",
    "streamlit>=1.52.2",
]
